Now handle the GPIO normally.
The GPIO package is now impossible to run outside of a raspberry pi as it loads the GPIO in memory as soon as it's loaded.

Any command you send to the GPIO through the given socket (/tmp/gpio.sock) will return a status update to every client connected to that socket.

The update payload take the form of a JSON response with the following fields:

{
  id: int
  status: boolean
}

The id field is the pin #.
The status field is a boolean that reflects the state of the GPIO (Low = 0, High = 1)